fix: allow fractional story points
Half points (0.5, 1.5) are normal estimation practice but three places
rejected them: the router's z.number().int(), the client's
Number.isInteger guard, and the input's step={1}.

Kept min(0)/max(1000) on the server rather than dropping to a bare
z.number(), which would accept Infinity. The input uses step="any" so the
browser never marks a value the server accepts as invalid.

+++ src/server/modules/ticketing/router.ts
@@ -19,7 +19,9 @@ const commentSchema = z.object({
 
 const adminUpdateSchema = z.object({
   stage: z.enum(customerStages as [CustomerStage, ...CustomerStage[]]).optional(),
-  storyPoints: z.number().int().min(0).max(1000).optional(),
+  // Half points (0.5, 1.5) are normal estimation practice, so not .int().
+  // Bounded rather than free-form: z.number() alone would accept Infinity.
+  storyPoints: z.number().min(0).max(1000).optional(),
   rawStatus: z.string().trim().min(1).optional(),
   title: z.string().trim().min(1).optional(),
   description: z.string().trim().optional(),
